Renamed packaging.py to packaging_metadata.py
This change was made to avoid naming conflicts when importing the packaging module from
packaging.py
