+++ checkbox-ng/pyproject.toml
@@ -66,7 +66,7 @@
   category = "plainbox.impl.unit.category:CategoryUnit"
   'test plan' = "plainbox.impl.unit.testplan:TestPlanUnit"
   'manifest entry' = "plainbox.impl.unit.manifest:ManifestEntryUnit"
-  'packaging meta-data' = "plainbox.impl.unit.packaging:PackagingMetaDataUnit"
+  'packaging meta-data' = "plainbox.impl.unit.packaging_metadata:PackagingMetaDataUnit"
   exporter = "plainbox.impl.unit.exporter:ExporterUnit"
 [project.entry-points."plainbox.parsers"]
   pxu = "plainbox.impl.secure.rfc822:load_rfc822_records"
