+++ checkbox-ng/setup.cfg
@@ -35,7 +35,7 @@ plainbox.unit=
   category=plainbox.impl.unit.category:CategoryUnit
   test plan=plainbox.impl.unit.testplan:TestPlanUnit
   manifest entry=plainbox.impl.unit.manifest:ManifestEntryUnit
-  packaging meta-data=plainbox.impl.unit.packaging:PackagingMetaDataUnit
+  packaging meta-data=plainbox.impl.unit.packaging_metadata:PackagingMetaDataUnit
   exporter=plainbox.impl.unit.exporter:ExporterUnit
 plainbox.parsers=
   pxu=plainbox.impl.secure.rfc822:load_rfc822_records
